pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarm-agency"
version = "1.0.0"
description = "43 AI personas across 10 departments — use as markdown rules or as a multi-model debate engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Miles0sage" }]
keywords = [
    "ai-agents",
    "multi-agent",
    "swarm-intelligence",
    "llm",
    "debate",
    "decision-making",
    "cursor-rules",
    "claude-code",
    "windsurf",
    "aider",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
cli = ["rich>=13.0.0"]
web = ["streamlit>=1.30.0"]
api = ["fastapi>=0.100.0", "uvicorn>=0.20.0", "sse-starlette>=1.6.0"]
dspy = ["dspy-ai>=2.5.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "rich>=13.0.0",
]

[project.scripts]
swarm-agency = "swarm_agency.cli:main"

[project.urls]
Homepage = "https://github.com/Miles0sage/swarm-agency"
Repository = "https://github.com/Miles0sage/swarm-agency"
Issues = "https://github.com/Miles0sage/swarm-agency/issues"
Documentation = "https://github.com/Miles0sage/swarm-agency#readme"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["swarm_agency"]
omit = ["tests/*"]
